Refactor ConsoleLog to be more generic by using message slot. Refactor uses of ConsoleLog to slot in custom message components.
